Revert "Feature/fvp forms"
