+++ nacc/uds3/ivp/packet.py
@@ -37,3 +37,4 @@ class Packet(list):
                 return self._cache[key]
 
         raise KeyError(key)
+
